pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "crostore"
description = "Allows you to sell your products over platforms"
readme = "README.md"
requires-python = "~=3.10"
license = "MIT"
authors = [
  { name = "Shuhei Nitta", email = "huisintheta@gmail.com" }
]
maintainers = [
  { name = "Shuhei Nitta", email = "huisintheta@gmail.com" }
]
# keywords = []
# classifiers = []
dependencies = [
  "selenium~=4.15.2",
]
dynamic = ["version"]

[project.urls]
Documentation = "https://github.com/huisint/crostore"
Homepage = "https://github.com/huisint/crostore"
Repository = "https://github.com/huisint/crostore"

[project.optional-dependencies]
test = [
  "mypy==1.9.0",
  "pytest==8.1.1",
  "pytest-cov==5.0.0",
  "pytest-describe==2.2.0",
  "pytest-mock==3.14.0",
]
google = [
  "google-api-python-client~=2.124.0",
  "google-api-python-client-stubs==1.25.0",
]

[tool.hatch.version]
path = "src/crostore/__init__.py"

[tool.hatch.envs.default]
features = ["test", "google"]

[tool.hatch.envs.default.scripts]
test = "pytest --cov-report=term-missing:skip-covered src tests {args}"
test-selenium = "test --selenium=$SELENIUM_URL {args}"
# test-doc = "pytest --doctest-modules src"
check-type = "mypy src tests {args}"

[tool.hatch.envs.lint]
skip-install = true
dependencies = [
  "black==23.11.0",
  "isort==5.12.0",
  "flake8==6.1.0",
]

[tool.hatch.envs.doc]
dependencies = []

[tool.hatch.envs.doc.scripts]

[tool.mypy]
strict = true
sqlite_cache = true

[[tool.mypy.overrides]]
module = [
  "google.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "--cov=crostore"

[tool.ruff]
lint.select = [
  "E",
  "F",
  "W",
  "I",
]
